requirements.txt
faiss-cpu==1.7.4
google-generativeai==0.2.2
langchain==0.0.325
openai==0.28.1
python-dotenv==1.0.0
streamlit==1.28.0
tiktoken==0.5.1
